Feature/cake nuget testing (#8928)
# Summary

Added an additional test/validation step in the build automation
workflow, the primary intention being to validate if projects can be
created and built prior to packages being deployed.

This builds projects using the following templates:

- DesktopGL
- WIndowsDX (only on Windows)
- Android
- iOS (only on Mac)
- Blank 2D Starter kit - larger scale project
- Full Starter kit - larger project with basic content

A report is generated with each run identifying the success state and
the versions of NuGet's used during testing, sourced from the local
build run.