Patch tests workflow for Windows (#9112)
# Summary

After applying environment and dotnet fixes for automation, windows
tests are now failing.

The issue was the yaml code to generate the global.json file was emitting invalid json on windows. 
The fix in this case is to make sure that on e windows we use `pwsh` as a shell rather than the default of `cmd`. 
The commit also makes sure we use `bash` on mac and linux.