Use HashSet to track GraphicsDevice resources (#9027)
<!--
Are you targeting develop? All PRs should target the develop branch
unless otherwise noted.
-->

Improves performance of `GraphicsDevice.RemoveResourceReference` (and by
extension `GraphicsResource.Dispose`) when there are a lot of resources.

<!--
Please try to make sure that there is a bug logged for the issue being
fixed if one is not present.
Especially for issues which are complex. 
The bug should describe the problem and how to reproduce it.
-->

### Description of Change

`List<T>.Remove(T item)` needs to iterate every item of the list to find
the one to remove, then shift items when removing from the middle.
HashSet<T> has much faster deletion, at the cost of a slightly slower
insert.

My specific example had over 30000 resources, and triggered large
stutters when clearing a cache with around 4000 recently allocated
items. It was iterating approximately 120000000 resources removing these
entries, since they would be at the end of the list.

My test case was a little extreme since I was scrolling a map with a
huge number of unique models that stream in+out, but it's an easy
improvement to make.

<!-- 
Enter description of the fix in this section.
Please be as descriptive as possible, future contributors will need to
know *why* these changes are being made.
For inspiration review the commit/PR history in the MonoGame repository.
-->